+++ Apizr/Samples/Apizr.Sample/Models/UserDetails.cs
@@ -1,10 +1,11 @@
 ﻿using System.Text.Json.Serialization;
+using Apizr.Configuring;
 using Apizr.Requesting;
 using Apizr.Requesting.Attributes;
 
 namespace Apizr.Sample.Models
 {
-    [CrudEntity("https://reqres.in/api/users")]
+    [BaseAddress("https://reqres.in/api/users")]
     public class UserDetails
     {
         [JsonPropertyName("data")]